[#177] refactor: 경로명 변경
