+++ src/main/java/com/beat/domain/performance/application/PerformanceUpdateService.java
@@ -6,6 +6,24 @@ import com.beat.domain.cast.exception.CastErrorCode;
 import com.beat.domain.member.dao.MemberRepository;
 import com.beat.domain.member.exception.MemberErrorCode;
 import com.beat.domain.performance.application.dto.update.*;
+import com.beat.domain.performance.application.dto.update.cast.CastAddRequest;
+import com.beat.domain.performance.application.dto.update.cast.CastAddResponse;
+import com.beat.domain.performance.application.dto.update.cast.CastDeleteRequest;
+import com.beat.domain.performance.application.dto.update.cast.CastDeleteResponse;
+import com.beat.domain.performance.application.dto.update.cast.CastUpdateRequest;
+import com.beat.domain.performance.application.dto.update.cast.CastUpdateResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleAddRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleAddResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleDeleteRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleDeleteResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleUpdateRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleUpdateResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffAddRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffAddResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffDeleteRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffDeleteResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffUpdateRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffUpdateResponse;
 import com.beat.domain.performance.dao.PerformanceRepository;
 import com.beat.domain.performance.domain.Performance;
 import com.beat.domain.performance.exception.PerformanceErrorCode;
