--- src/main/java/com/beat/domain/performance/application/dto/update/CastDeleteRequest.java
@@ -1,5 +1,0 @@
-package com.beat.domain.performance.application.dto.update;
-
-public record CastDeleteRequest(
-        Long castId
-) {}
