+++ src/main/java/com/beat/domain/performance/application/dto/update/PerformanceUpdateRequest.java
@@ -1,5 +1,14 @@
 package com.beat.domain.performance.application.dto.update;
 
+import com.beat.domain.performance.application.dto.update.cast.CastAddRequest;
+import com.beat.domain.performance.application.dto.update.cast.CastDeleteRequest;
+import com.beat.domain.performance.application.dto.update.cast.CastUpdateRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleAddRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleDeleteRequest;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleUpdateRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffAddRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffDeleteRequest;
+import com.beat.domain.performance.application.dto.update.staff.StaffUpdateRequest;
 import com.beat.domain.performance.domain.BankName;
 import com.beat.domain.performance.domain.Genre;
