+++ src/main/java/com/beat/domain/performance/application/dto/update/PerformanceUpdateResponse.java
@@ -1,5 +1,14 @@
 package com.beat.domain.performance.application.dto.update;
 
+import com.beat.domain.performance.application.dto.update.cast.CastAddResponse;
+import com.beat.domain.performance.application.dto.update.cast.CastDeleteResponse;
+import com.beat.domain.performance.application.dto.update.cast.CastUpdateResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleAddResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleDeleteResponse;
+import com.beat.domain.performance.application.dto.update.schedule.ScheduleUpdateResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffAddResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffDeleteResponse;
+import com.beat.domain.performance.application.dto.update.staff.StaffUpdateResponse;
 import com.beat.domain.performance.domain.BankName;
 import com.beat.domain.performance.domain.Genre;
