--- src/main/java/com/beat/domain/performance/application/dto/update/ScheduleDeleteRequest.java
@@ -1,5 +1,0 @@
-package com.beat.domain.performance.application.dto.update;
-
-public record ScheduleDeleteRequest(
-        Long scheduleId
-) {}
