--- src/main/java/com/beat/domain/performance/application/dto/update/StaffDeleteRequest.java
@@ -1,5 +1,0 @@
-package com.beat.domain.performance.application.dto.update;
-
-public record StaffDeleteRequest(
-        Long staffId
-) {}
