+++ src/main/java/com/beat/domain/performance/application/dto/update/cast/CastDeleteRequest.java
@@ -1,0 +1,5 @@
+package com.beat.domain.performance.application.dto.update.cast;
+
+public record CastDeleteRequest(
+        Long castId
+) {}
