+++ src/main/java/com/beat/domain/performance/application/dto/update/cast/CastDeleteResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.cast;
 
 public record CastDeleteResponse(
         Long castId
