+++ src/main/java/com/beat/domain/performance/application/dto/update/cast/CastUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.cast;
 
 public record CastUpdateResponse(
         Long castId,
