+++ src/main/java/com/beat/domain/performance/application/dto/update/schedule/ScheduleAddRequest.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.schedule;
 
 import com.beat.domain.schedule.domain.ScheduleNumber;
