+++ src/main/java/com/beat/domain/performance/application/dto/update/schedule/ScheduleDeleteRequest.java
@@ -1,0 +1,5 @@
+package com.beat.domain.performance.application.dto.update.schedule;
+
+public record ScheduleDeleteRequest(
+        Long scheduleId
+) {}
