+++ src/main/java/com/beat/domain/performance/application/dto/update/schedule/ScheduleDeleteResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.schedule;
 
 public record ScheduleDeleteResponse(
         Long scheduleId
