+++ src/main/java/com/beat/domain/performance/application/dto/update/staff/StaffAddRequest.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.staff;
 
 public record StaffAddRequest(
         String staffName,
