+++ src/main/java/com/beat/domain/performance/application/dto/update/staff/StaffDeleteRequest.java
@@ -1,0 +1,5 @@
+package com.beat.domain.performance.application.dto.update.staff;
+
+public record StaffDeleteRequest(
+        Long staffId
+) {}
