+++ src/main/java/com/beat/domain/performance/application/dto/update/staff/StaffDeleteResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.performance.application.dto.update;
+package com.beat.domain.performance.application.dto.update.staff;
 
 public record StaffDeleteResponse(
         Long staffId
